requirements.txt
pandas
streamlit
groq
beautifulsoup4
python-dotenv
duckduckgo-search
phidata
google-cloud-firestore